Add files via upload
This assignment included writing a complete C++ program that does the following:
o Prompt the user for their first name and last name (store in fName and
lName variables, respectively)
o Pass the two strings to a function called swapFirstLetters that modifies
each string by replacing its first letter with that of the other.
  - For example, if fName = Dave and lName = Miller were passed, the
strings would be modified by the function to store fName = Mave
and lName = Diller.
o Print fName and lName to the monitor.
o Pass the two strings to a function called longestString that prints the larger
string and returns the length of the larger string. Store this return value in
an int variable called length.
- If both strings are of equal length, you can print whichever string
you prefer.
o Pass length to a function called repeatHello that prints "Hello" to the
monitor length times.